requirements.txt
anyio==4.9.0
certifi==2025.4.26
et_xmlfile==2.0.0
flet==0.28.2
flet-desktop==0.28.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
repath==0.9.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
